Start Markdown documentation of the maps system
This starts a human-readable documentation of the various subsystems
used in the game's map system. The following files are included in this
commit:

- `maps.md`: a general overview of maps
- `bdhc.md`: an overview of the BDHC subsystem, which is used to provide
  height information on maps
- `loading_maps.md`: an overview of how maps are loaded by the game
- `file_format_specifications.md`: specs for the maps data structures

The `file_format_specifications.md` is meant to be a purely technical
document that contains a precise description of the data structures used
by the game's map system.

The other files are meant to be semi-technical documentation, where some
of the nitty-gritty details have been either simplified or left out
entirely.

Some of the maps subsystems were purposefully left out of this commit,
such as map objects and map scripts. They will need to be addressed in
future updates to those documents.

Signed-off-by: Kuruyia <github@kuruyia.net>