Document distortion world floating platforms
This continues documentation of the distortion world overlay by
documenting the contents of the first section of files in the
`tw_arc.narc` archive.

This section contains the data for floating platforms contained within
one of the maps that compose the distortion world.

Here, a floating platform is defined as the platforms that the player
can walk on and that changes their orientation. This _does not_ take
into account the platforms that the player predominantly walks on during
their time in the distortion world, on which the player keeps the
regular orientation.

This also documents code related to loading the terrain attributes
for each floating platform from the `tw_arc_attr.narc` archive.

Finally, this also documents functions that help get collision and tile
attributes from the floating platform the player is currently standing
on.

Signed-off-by: Kuruyia <github@kuruyia.net>